journal-dev-1 | Child Component Separation
For cleaner code, I separated the actual journal entry editor, and the log in components into child components.

This way the "journal editor" main parent component looks cleaner with less code and the actual function is hidden within the child components respectively.

This also helps with performance as it won't be necessary to reload two components on an entry change/password entry.

I just need to confirm if the journal entry still works as intended.
